Inline plan details service and formatter
The PlanDetailsService and PlanDetailsFormatter each had a single caller after unifying the plan details use case. Fold their logic into GetPlanDetailsInteractor and GetPlanDetailsPresenter respectively.
