+++ tests/interactors/test_file_plan_with_accounting.py
@@ -7,12 +7,14 @@ from workers_control.core import email_notifications
 from workers_control.core.interactors.file_plan_with_accounting import (
     FilePlanWithAccounting,
 )
-from workers_control.core.interactors.get_plan_details import GetPlanDetailsInteractor
+from workers_control.core.interactors.get_plan_details import (
+    GetPlanDetailsInteractor,
+    PlanDetails,
+)
 from workers_control.core.interactors.list_plans_with_pending_review import (
     ListPlansWithPendingReviewInteractor,
 )
 from workers_control.core.records import ProductionCosts
-from workers_control.core.services.plan_details import PlanDetails
 
 from .base_test_case import BaseTestCase
